WIN32 fixes
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@17079 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3944fc1e31fb5527b57183f4ea473e2fcd7eb6e2

+++ LHCb/Tf/TfKernel/TfKernel/DefaultVeloHitManager.h
@@ -1,4 +1,4 @@
-// $Id: DefaultVeloHitManager.h,v 1.7 2007-08-26 14:10:41 krinnert Exp $
+// $Id: DefaultVeloHitManager.h,v 1.8 2007-08-28 12:03:58 jonrob Exp $
 #ifndef INCLUDE_TF_DEFAULTVELOHITMANAGER_H
 #define INCLUDE_TF_DEFAULTVELOHITMANAGER_H 1
 
@@ -129,7 +129,7 @@ namespace Tf {
 
       for(iClus = clusters->begin(); iClus != clusters->end(); ++iClus) {
 
-        float signal = (*iClus)->totalCharge();
+        double signal = (*iClus)->totalCharge();
 
         // don't use this cluster if we run in ingore mode and the total charge is too low
         if ( m_ignoreLowClusters && (signal < m_lowThreshold) ) continue;
